Add wire format parsing unit tests for implicit presence fields.
This gist of the new test case coverage can be summarized as:
- The wire does not distinguish between explicit and implicit fields.
- For implicit presence fields, zeros on the wire can overwrite nonzero values
  (i.e. they are treated as a 'clear' operation).

It's TBD whether we want to accept this behaviour going forward. It might be
somewhat problematic if we have nondeterministic field ordering. Or, we might
just want to call this unsafe behaviour (well-behaved code should never
serialize the same field multiple times on the wire).

I still need to add conformance tests to ensure that this is a consistent
behavior across all languages, but for now let's at least add some coverage in
C++ to ensure that this is a tested behaviour.

PiperOrigin-RevId: 656535937