Rust: introduce trait for access to message default instances
The immediate motivation for this is that it will facilitate writing a blanket
implementation of `ProxiedInMapValue` for C++-backed messages. The default
instance gives us access to the message vtable in cases where we don't already
have a message to work with.

It also seems useful for code that just needs a static-lifetime empty view,
though, so I implemented it for upb messages as well.

PiperOrigin-RevId: 676509188