Introduce set_alias to hpb
In certain cases, it is useful to share submessages across multiple parent messages.
proto2::cpp has a mechanism for this, so we add the hpb equivalent.

For this initial impl, we stipulate that the arenas must be exactly the same. We may explore broadening the constraint to allow for all fused arenas.

Remark: if the arenas do not match, set_alias will explode at runtime, so beware (and check before calling).
PiperOrigin-RevId: 680695598