Ruby: fix bug in Map.hash
`Map.hash` currently has a bug, which is that it assumes a consistent iteration
order. Experimentally it appears that the iteration order is usually
consistent, but not when there is a collision in the hash table. This can lead
to two equal maps producing different hash values.

This change fixes the problem by making `Map.hash` add up the hash values of
each of its elements, instead of feeding them sequentially into upb's hash
function. Since addition is commutative, this produces a consistent result
regardless of iteration order.

PiperOrigin-RevId: 723526472