Moved and split fasttable decoder into the `decode_fast` directory.
This CL contains almost no functional change.  It is mainly a splitting that prepares for subsequent changes.  However we do remove the dependency edge from select.cc to reflection, and this results in a few changes to `select.cc` as we make it based on MiniTables instead of reflection.

In particular, we no longer use a fields required property as a proxy for hotness information.  This would technically be possible to do, as MiniTables do know this information, but there is no function for it currently and I'm reluctant to add one as we strongly discourage the use of required fields.

We also remove a check for message decoding that the sub-message is in the current file.  In the future we can fix this by adding a cheap size check into the decoding path.

Splitting `decode_fast.c` into multiple files makes the code easier to navigate and understand.

PiperOrigin-RevId: 752060149