Work around issues with C++20 in Clang 15 and older
Older versions of Clang seem to have trouble with `absl::flat_hash_map` when we
build for C++20 if the value type is not fully defined. For some reason this
works today, but stops working if we upgrade Abseil to the latest version.

This is particularly a problem for `UntypedMessage::fields_` in
`untyped_message.h`, so this changes works around that problem by wrapping the
value in a `std::unique_ptr`.

PiperOrigin-RevId: 773090861