Add a DCHECK that ArenaString::Set(char*, Arena*) is not called with a null char* argument.
This is already undefined behavior since the argument is immediately
used to construct a string_view, and passing a null argument to the
single-arg string_view constructor is undefined behavior. Adding the
check will prevent introduction of new bugs when using standard
library implementations that do not detect this case.

PiperOrigin-RevId: 781510912