Remove the flag requirement for Bazel+MSVC users.
As discussed in https://github.com/protocolbuffers/protobuf/issues/20085, we were originally intending to drop support for this combination.  With Bazel's recent changes to virtual includes on windows though, we now have ~90 extra characters of runway.  This should be more than enough to continue supporting MSVC into the forseeable future.  Clang-cl support will be kept in place as an alternative on windows

PiperOrigin-RevId: 783391994