Split `Arena::Construct` into two methods, one which assumes we are using arenas.
This allows us to avoid a conditional in `DoCreateMessage`, which is only called when we are using arenas.

Note that we also remove the logic for constructing a wrapper type around the allocated `T`, as this logic was erroneous (and also never triggered). If a `T` was allocated, we only allocated enough memory for `T`, so constructing a `typename internal::FieldArenaRep<T>::Type` is UB. This never triggers because all callers of `Construct` would have already translated `T` to its arena rep, and the arena rep of an arena rep is always itself.

PiperOrigin-RevId: 810313180