Fix `Any` hasbit consistency issue in OSS.
`Any` `value` field is a `bytes` field in OSS, and since cl/792909628 (https://github.com/protocolbuffers/protobuf/pull/22956), `_internal_mutable_*()` accessors for string fields don't set hasbits. This can cause the hasbit to be missing for the `value` field of `Any` after calling `PackFrom`, which will serialize incorrectly.

The culprit change was adopted in release 33.0, so it will need to be patched to this version.

See https://github.com/protocolbuffers/protobuf/issues/24258.

PiperOrigin-RevId: 828033634